+++ src/components/modal/UserSettingsForm/UserSettingsForm.module.css
@@ -91,10 +91,6 @@
   }
 }
 
-.fileInput {
-  display: none;
-}
-
 .uploadIcon {
   width: 18px;
   height: 18px;
